# speed up nearest neighbor query by pruning